+++ colander/__init__.py
@@ -1802,8 +1802,9 @@ class DateTime(SchemaType):
 
     err_template = _('Invalid date')
 
-    def __init__(self, default_tzinfo=iso8601.UTC):
+    def __init__(self, default_tzinfo=iso8601.UTC, format=None):
         self.default_tzinfo = default_tzinfo
+        self.format = format
 
     def serialize(self, node, appstruct):
         if not appstruct:
@@ -1824,17 +1825,25 @@ class DateTime(SchemaType):
 
         if appstruct.tzinfo is None:
             appstruct = appstruct.replace(tzinfo=self.default_tzinfo)
-        return appstruct.isoformat()
+        if not self.format:
+            return appstruct.isoformat()
+        else:
+            return appstruct.strftime(self.format)
 
     def deserialize(self, node, cstruct):
         if not cstruct:
             return null
 
         try:
-            result = iso8601.parse_date(
-                cstruct, default_timezone=self.default_tzinfo
+            if self.format:
+                result = datetime.datetime.strptime(cstruct, self.format)
+                if not result.tzinfo and self.default_tzinfo:
+                    result = result.replace(tzinfo=self.default_tzinfo)
+            else:
+                result = iso8601.parse_date(
+                    cstruct, default_timezone=self.default_tzinfo
             )
-        except iso8601.ParseError as e:
+        except (ValueError, iso8601.ParseError) as e:
             raise Invalid(
                 node, _(self.err_template, mapping={'val': cstruct, 'err': e})
             )
@@ -1883,6 +1892,9 @@ class Date(SchemaType):
 
     err_template = _('Invalid date')
 
+    def __init__(self, format=None):
+        self.format = format
+
     def serialize(self, node, appstruct):
         if not appstruct:
             return null
@@ -1896,13 +1908,18 @@ class Date(SchemaType):
                 _('"${val}" is not a date object', mapping={'val': appstruct}),
             )
 
+        if self.format:
+            return appstruct.strftime(self.format)
         return appstruct.isoformat()
 
     def deserialize(self, node, cstruct):
         if not cstruct:
             return null
         try:
-            result = iso8601.parse_date(cstruct)
+            if self.format:
+                result = datetime.datetime.strptime(cstruct, self.format)
+            else:
+                result = iso8601.parse_date(cstruct)
             result = result.date()
         except iso8601.ParseError as e:
             raise Invalid(